work around to prevent stack underflowing
compiler emits OP_RETURN opcode when done which with the current implementation
pops the stack causing an underflow because statements don't leave any values
on the stack.
with this commit statements leave their result on the stack and OP_RETURN checks
the stack properly and print what's in it.
this behaviour should be removed when possible.